Create method to delete files after delete in admin site

+++ funbox/media/models.py
@@ -84,17 +84,17 @@ class Soundtrack(Media):
         return self.config__str__('soundtrack')
 
 @receiver(models.signals.post_delete, sender=Soundtrack)
-def auto_delete_file_on_delete(sender, instance, **kwargs):
+def auto_delete_soundtrack_on_delete(sender, instance, **kwargs):
     """
     Deletes file from filesystem
-    when corresponding `Media` object is deleted.
+    when corresponding `Soundtrack` object is deleted.
     """
     if instance.soundtrack:
         if os.path.isfile(instance.soundtrack.path):
             os.remove(instance.soundtrack.path)
 
 @receiver(models.signals.pre_save, sender=Soundtrack)
-def auto_delete_file_on_change(sender, instance, **kwargs):
+def auto_delete_soundtrack_on_change(sender, instance, **kwargs):
     """
     Deletes old file from filesystem
     when corresponding `Soundtrack` object is updated
@@ -109,6 +109,51 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
         return False
 
     new_file = instance.soundtrack
+    if not old_file == new_file:
+        if os.path.isfile(old_file.path):
+            os.remove(old_file.path)
+
+@receiver(models.signals.post_delete, sender=Image)
+def auto_delete_image_on_delete(sender, instance, **kwargs):
+    if instance.image:
+        if os.path.isfile(instance.image.path):
+            os.remove(instance.image.path)
+
+@receiver(models.signals.pre_save, sender=Image)
+def auto_delete_image_on_change(sender, instance, **kwargs):
+    """
+    Deletes old file from filesystem
+    when corresponding `Image` object is updated
+    with new file.
+    """
+    if not instance.pk:
+        return False
+
+    try:
+        old_file = Image.objects.get(pk=instance.pk).image
+    except Image.DoesNotExist:
+        return False
+
+    new_file = instance.image
+    if not old_file == new_file:
+        if os.path.isfile(old_file.path):
+            os.remove(old_file.path)
+
+@receiver(models.signals.post_delete, sender=Video)
+def auto_delete_video_on_delete(sender, instance, **kwargs):
+    if instance.video:
+        if os.path.isfile(instance.video.path):
+            os.remove(instance.video.path)
+
+@receiver(models.signals.pre_save, sender=Video)
+def auto_delete_video_on_change(sender, instance, **kwargs):
+    if not instance.pk:
+        return False
+    try:
+        old_file = Video.objects.get(pk=instance.pk).video
+    except Video.DoesNotExist:
+        return False
+    new_file = instance.video
     if not old_file == new_file:
         if os.path.isfile(old_file.path):
             os.remove(old_file.path)
